requirements.txt
# AI Agent Trader — Python dependencies
# Install: pip install -r requirements.txt

# LangGraph & LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# API & Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# UI
streamlit>=1.40.0
requests>=2.32.0

# Data & Finance
yfinance>=0.2.0
pandas>=2.0.0
finnhub-python>=1.4.0
stockstats>=0.7.0

# Memory & Embeddings
chromadb>=0.5.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
typing_extensions>=4.0.0
